refactor(button): :fire: move button props type to the component file itself
